docker_compose/app2/requirements.txt
flask==2.3.3
pyzmq==25.1.2